Fix unlock timer issue
The variable unlock_duration.seconds is an integer so when you convert it to an hour by dividing by 3600 it returns an integer, which is often 0.  So I have converted it to a float before doing the division.  Hopefully this fixes this sort of issues, but there may be other when different duration timers are used.