Use state_long from geocoder, negating need for states dict.
Geocoder has a property called state_long.  This returns the full state or province name in long form, ie as New Hampshire rather than NH.  I have tried it for the US, Canada and Australia so far.  We can always expand the list from these countries if we find that it works for my countries.